% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_pal}
\alias{get_pal}
\title{Get a Pantone 35 color palette as a character vector}
\usage{
get_pal(palette_name)
}
\arguments{
\item{palette_name}{the name of the palette}
}
\value{
a character vector of length 4
}
\description{
Get a Pantone 35 color palette as a character vector
}
\examples{
get_pal("ivy league")

}
