% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{List of Pantone color palettes.}
\format{An object of class \code{list} of length 35.}
\usage{
palettes
}
\description{
List of Pantone color palettes.
}
\keyword{datasets}
